requirements.txt
openai==2.16.0
tiktoken==0.12.0
faiss-cpu==1.13.2
rank-bm25==0.2.2
numpy==2.2.2
pydantic==2.12.5
pytest==9.0.2
